requirements.txt
streamlit==1.22.0
python-telegram-bot==20.3